forward port of 93bb95929a0c017498625a87654b4d114fe17b96 to allow mounting and sharing of mulitple volumes to MountService.
Change-Id: I9fa3994e6e7b85360b451f57c1cdd4df921bc035

Wait for IDLE state before mounting on insert

As soon as we recive the insert event we try to
mount the mmc sending a "volume mount" command.
But, the volume manager sometimes is still on
"pending" state, and the insertion mount will
fail.

This patch calls the doMount command on insert
only after the state has changed to idle.
This way, the mount will not fail, because the
state is pending.

Change-Id: I1f8f13c8e232a6b9159eeb78a9484838f1c600af
Signed-off-by: Axel Haslam <axelhaslam@ti.com>