Add BOARD_FIRST_CAMERA_FRONT_FACING betelgeuse (folio100) support
betelgeuse has only one front facing camera in landscape. Add define to
allow this configuration.
This defines allow to modify the first CameraInfo parameters of the
sCameraInfo array.
The modification of CamcorderProfile is to avoid NPE in Camera app,
because frameworks assume that only back facing camera can record video.

Change-Id: I071c5bdcf55c722c05295299b34082f4a2ea01eb