More ergonomic TimePicker and DatePicker
Currently, when the text in the TimePicker or in the DatePicker dialog
is needed to change, it is necessary to either click several times on
the spinner buttons or manually rewrite the values with
software/hardware keyboard. The manual rewrite is very slow as it is
necessary to delete the text before new one can be entered. This change
implements more ergonomic behavior of the TimePicker and the DatePicker
which allow to enter time or date much faster. If the editable field is
focused, all text is automatically selected in order to be ready for
rewrite. If it is not the last field in the row and if the maximum
length of text is reached, the next field is automatically selected.

In the case of TimePicker:
If the Hour field is focused, all text is selected and ready for
rewrite. Once two characters are entered, the Minute field is focused.
When the Minute filed is focused, all text is selected and ready for
rewrite.

In the case of DatePicker (US locale):
If the Month field is focused, all text is selected and ready for
rewrite. Once three characters are entered, the Day field is focused.
When the Day field is focused, all text is selected and ready for
rewrite. Once two characters are entered, the Year field is focused.
When the Year field is focused, all text is selected and ready for
rewrite. The order of the fields vary according to user's locale.

Change-Id: Ibe5d678c80f2803d698340aa0c8f795cdc7a7e03